fix: address fade curve review
