+++ crates/vibez-ui/src/widgets/timeline/fade_drag.rs
@@ -17,6 +17,7 @@ pub struct FadeClipDrag {
     clip_x: f32,
     clip_width: f32,
     duration_frames: u64,
+    handle: Point,
 }
 
 #[derive(Debug, Clone)]
@@ -26,17 +27,28 @@ pub struct FadeCurveDrag {
     pub edge: AudioClipFadeEdge,
     top: f32,
     bottom: f32,
+    handle: Point,
+}
+
+pub(super) enum FadeControlDrag {
+    Length(FadeClipDrag),
+    Curve(FadeCurveDrag),
 }
 
 impl FadeCurveDrag {
-    fn new(clip_id: ClipId, edge: AudioClipFadeEdge, canvas_height: f32) -> Self {
+    fn new(clip_id: ClipId, edge: AudioClipFadeEdge, canvas_height: f32, handle: Point) -> Self {
         Self {
             undo_gesture: UndoGestureId::new(),
             clip_id,
             edge,
             top: CLIP_Y + CLIP_TITLE_HEIGHT + 2.0,
             bottom: canvas_height - CLIP_Y - 2.0,
+            handle,
         }
+    }
+
+    fn distance_squared(&self, point: Point) -> f32 {
+        (point.x - self.handle.x).powi(2) + (point.y - self.handle.y).powi(2)
     }
 
     pub(super) fn curve_at_y(&self, y: f32) -> FadeCurve {
@@ -57,6 +69,7 @@ impl FadeClipDrag {
         clip_x: f32,
         clip_width: f32,
         duration_frames: u64,
+        handle: Point,
     ) -> Self {
         Self {
             undo_gesture: UndoGestureId::new(),
@@ -65,7 +78,12 @@ impl FadeClipDrag {
             clip_x,
             clip_width,
             duration_frames,
+            handle,
         }
+    }
+
+    fn distance_squared(&self, point: Point) -> f32 {
+        (point.x - self.handle.x).powi(2) + (point.y - self.handle.y).powi(2)
     }
 
     /// Use the exact inverse of [`fade_handle_xs`] so a handle drawn at one
@@ -139,7 +157,12 @@ impl TrackClipCanvas {
                 if (pos.x - handle.x).abs() <= FADE_HANDLE_HIT_RADIUS
                     && (pos.y - handle.y).abs() <= FADE_HANDLE_HIT_RADIUS
                 {
-                    return Some(FadeCurveDrag::new(clip.clip_id, edge, canvas_height));
+                    return Some(FadeCurveDrag::new(
+                        clip.clip_id,
+                        edge,
+                        canvas_height,
+                        handle,
+                    ));
                 }
             }
         }
@@ -175,16 +198,46 @@ impl TrackClipCanvas {
                 None
             };
             if let Some(edge) = edge {
+                let handle_x = match edge {
+                    AudioClipFadeEdge::In => fade_in_x,
+                    AudioClipFadeEdge::Out => fade_out_x,
+                };
                 return Some(FadeClipDrag::new(
                     clip.clip_id,
                     edge,
                     clip_x,
                     clip_width,
                     clip.duration,
+                    Point::new(handle_x, FADE_HANDLE_Y),
                 ));
             }
         }
         None
+    }
+
+    /// Resolve overlapping fade controls by their drawn distance. The
+    /// fade-length handle wins an exact tie so a short steep fade can always
+    /// be made longer again.
+    pub(super) fn fade_control_hit(
+        &self,
+        pos: Point,
+        canvas_height: f32,
+    ) -> Option<FadeControlDrag> {
+        match (
+            self.fade_handle_hit(pos),
+            self.fade_curve_hit(pos, canvas_height),
+        ) {
+            (Some(length), Some(curve)) => {
+                if length.distance_squared(pos) <= curve.distance_squared(pos) {
+                    Some(FadeControlDrag::Length(length))
+                } else {
+                    Some(FadeControlDrag::Curve(curve))
+                }
+            }
+            (Some(length), None) => Some(FadeControlDrag::Length(length)),
+            (None, Some(curve)) => Some(FadeControlDrag::Curve(curve)),
+            (None, None) => None,
+        }
     }
 }
 
@@ -220,6 +273,7 @@ mod tests {
             37.0,
             481.0,
             clip.duration,
+            Point::new(fade_in_x, FADE_HANDLE_Y),
         );
         let fade_out = FadeClipDrag::new(
             clip.clip_id,
@@ -227,6 +281,7 @@ mod tests {
             37.0,
             481.0,
             clip.duration,
+            Point::new(fade_out_x, FADE_HANDLE_Y),
         );
 
         assert_eq!(fade_in.frames_at_x(fade_in_x), clip.fade_in_frames);
